Separate amount definition from feerate
