+++ divi/src/FeeAndPriorityCalculator.cpp
@@ -1,5 +1,7 @@
 #include <FeeAndPriorityCalculator.h>
+
 #include "primitives/transaction.h"
+#include "FeeRate.h"
 
 CFeeRate minRelayTxFee = CFeeRate(10000);
